requirements.txt
requests
pydub
pympi-ling
